+++ AeroDms/PdfRenderer.h
@@ -73,7 +73,7 @@ private :
 
 	AeroDmsTypes::EtatGeneration imprimerLaProchaineDemandeDeSubvention();
 	void remplirLeChampMontant( QString& p_html, 
-		                        const float p_montant) const;
+		                        const double p_montant) const;
 	void remplirLeChampSignature(QString& p_html) const;
 
 	AeroDmsTypes::EtatGeneration imprimerLeFichierPdfDeRecapAnnuel( const int p_annee,
